requirements.txt
fastapi
uvicorn[standard]
python-multipart
python-dotenv
tensorflow
numpy
Pillow
